semistandard

+++ 0x12-javascript-warm_up/7-multi_c.js
@@ -4,7 +4,7 @@
 let i = 0;
 const myString = 'C is fun';
 
-if (Number(process.argv[2]) === parseFloat(process.argv[2])) {
+if (Number(process.argv[2]) === parseInt(process.argv[2])) {
   while (i < process.argv[2]) {
     console.log(myString);
     i++;
